refactor(delete-hole): pick hole and delete in single step
